requirements.txt
streamlit
langchain
langchain-community
chromadb
sentence-transformers
google-generativeai
